pcli: start splitting up main.rs into modules
