+++ pcli/src/warning.rs
@@ -1,0 +1,19 @@
+pub fn display() {
+    println!(
+        "
+                               \x1b[1;31m⛔️WARNING️️⛔️:
+        
+                        you are about to lose money!\x1b[0m
+
+🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥
+This message ... is part of a system of messages...
+... we considered ourselves to be a fault-tolerant distributed system...
+This message is a warning about danger.
+The danger is in a particular location... the center of danger is the \x1b[1;92mpcli\x1b[0m binary...
+The danger is to your \x1b[1;92mfunds\x1b[0m, and it can \x1b[1;31mdestroy\x1b[0m them.
+The danger is unleashed only if you execute this software.
+🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥
+
+                           \x1b[1;31mUSE AT YOUR OWN RISK\x1b[0m"
+    )
+}
